Added in a number constant for the numbered output flows needed for any flow node or otherwise; I only went up to 20, so any flow past [19] will be invalid.
Wrote a spec and test for the sequence node. Simply prints what flow is being run out of the sequence node. There's a quick-and-dirty value swap to ensure the sequence node is executing the output flows in the correct order; the test spits a log at the start of which values are swapped.